refactor into 2 files - python version
